% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dlm_operations.R
\name{delete_lifecycle_policy}
\alias{delete_lifecycle_policy}
\title{Deletes the specified lifecycle policy and halts the automated operations that the policy specified}
\usage{
delete_lifecycle_policy(PolicyId)
}
\arguments{
\item{PolicyId}{The identifier of the lifecycle policy.}
}
\description{
Deletes the specified lifecycle policy and halts the automated operations that the policy specified.
}
\examples{

}
