% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dlm_operations.R
\name{get_lifecycle_policy}
\alias{get_lifecycle_policy}
\title{Gets detailed information about the specified lifecycle policy}
\usage{
get_lifecycle_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The identifier of the lifecycle policy.}
}
\description{
Gets detailed information about the specified lifecycle policy.
}
\section{Accepted Parameters}{
\preformatted{get_lifecycle_policy(
  PolicyId = "string"
)
}
}

