% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dlm_package.R
\docType{package}
\name{paws.dlm-package}
\alias{paws.dlm}
\alias{paws.dlm-package}
\title{paws.dlm: Amazon Data Lifecycle Manager}
\description{
With Amazon Data Lifecycle Manager, you can manage
    the lifecycle of your AWS resources. You create lifecycle policies,
    which are used to automate operations on the specified resources.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
